move the definiton of pic_checkpoint to object.h
